+++ lib/src/model/temtem_api_cosmetic.dart
@@ -21,13 +21,13 @@ class TemTemApiCosmetic {
 
   factory TemTemApiCosmetic.fromJson(Map<String, dynamic> json) =>
       TemTemApiCosmetic(
-        type: json['type'],
-        wikiImageUrl: json['wikiImageUrl'],
-        wikiUrl: json['wikiUrl'],
-        name: json['name'],
-        location: json['location'],
-        cost: json['cost'],
-        requirement: json['requirement'],
-        description: json['description'],
+        type: json['type'] as String,
+        wikiImageUrl: json['wikiImageUrl'] as String,
+        wikiUrl: json['wikiUrl'] as String,
+        name: json['name'] as String,
+        location: json['location'] as String,
+        cost: json['cost'] as int,
+        requirement: json['requirement'] as String,
+        description: json['description'] as String,
       );
 }