+++ lib/src/model/temtem_api_location.dart
@@ -10,9 +10,9 @@ class Location {
   });
 
   factory Location.fromJson(Map<String, dynamic> json) => Location(
-        name: json['name'],
-        wikiUrl: json['wikiUrl'],
-        type: json['type'],
+        name: json['name'] as String,
+        wikiUrl: json['wikiUrl'] as String,
+        type: json['type'] as String,
       );
 }
 
@@ -24,9 +24,9 @@ class TemTemApiLocation {
   final String type;
   final String imageWikiThumbnail;
   final String imageWikiFile;
-  final List<Location> routes;
-  final List<Location> townsAndVillages;
-  final List<Location> landmarks;
+  final List<Location>? routes;
+  final List<Location>? townsAndVillages;
+  final List<Location>? landmarks;
   final List<String> temtem;
   final List<String> trivia;
 
@@ -45,27 +45,25 @@ class TemTemApiLocation {
     required this.trivia,
   });
 
-  factory TemTemApiLocation.fromJson(Map<String, dynamic> json) =>
-      TemTemApiLocation(
-        name: json['name'],
-        wikiUrl: json['wikiUrl'],
-        description: json['description'],
-        temtemTypes: List<String>.from(json['temtemTypes']),
-        type: json['type'],
-        imageWikiThumbnail: json['imageWikiThumbnail'],
-        imageWikiFile: json['imageWikiFile'],
-        routes: List<Location>.generate(
-            json['routes'] == null ? 0 : json['routes'].length,
-            (index) => Location.fromJson(json['routes'][index])),
-        townsAndVillages: List<Location>.generate(
-            json['townsAndVillages'] == null
-                ? 0
-                : json['townsAndVillages'].length,
-            (index) => Location.fromJson(json['townsAndVillages'][index])),
-        landmarks: List<Location>.generate(
-            json['landmarks'] == null ? 0 : json['landmarks'].length,
-            (index) => Location.fromJson(json['landmarks'][index])),
-        temtem: List<String>.from(json['temtem']),
-        trivia: List<String>.from(json['trivia']),
-      );
+  factory TemTemApiLocation.fromJson(Map<String, dynamic> json) {
+    final routes = (json['routes'] as Iterable?)?.cast<Map<String, dynamic>>();
+    final townsAndVillages =
+        (json['townsAndVillages'] as Iterable?)?.cast<Map<String, dynamic>>();
+    final landmarks =
+        (json['landmarks'] as Iterable?)?.cast<Map<String, dynamic>>();
+    return TemTemApiLocation(
+      name: json['name'] as String,
+      wikiUrl: json['wikiUrl'] as String,
+      description: json['description'] as String,
+      temtemTypes: List<String>.from(json['temtemTypes'] as Iterable),
+      type: json['type'] as String,
+      imageWikiThumbnail: json['imageWikiThumbnail'] as String,
+      imageWikiFile: json['imageWikiFile'] as String,
+      routes: routes?.map(Location.fromJson).toList(),
+      townsAndVillages: townsAndVillages?.map(Location.fromJson).toList(),
+      landmarks: landmarks?.map(Location.fromJson).toList(),
+      temtem: List<String>.from(json['temtem'] as Iterable),
+      trivia: List<String>.from(json['trivia'] as Iterable),
+    );
+  }
 }