+++ lib/src/model/temtem_api_technique.dart
@@ -3,12 +3,16 @@ class SynergyEffect {
   final String type;
   final int damage;
 
-  SynergyEffect({this.effect, this.type, this.damage});
+  SynergyEffect({
+    required this.effect,
+    required this.type,
+    required this.damage,
+  });
 
   factory SynergyEffect.fromJson(Map<String, dynamic> json) => SynergyEffect(
-        effect: json['effect'],
-        type: json['type'],
-        damage: json['damage'],
+        effect: json['effect'] as String,
+        type: json['type'] as String,
+        damage: json['damage'] as int,
       );
 }
 
@@ -29,37 +33,38 @@ class TemTemApiTechnique {
   final String description;
 
   TemTemApiTechnique({
-    this.name,
-    this.wikiUrl,
-    this.type,
-    this.classTouch,
-    this.classIcon,
-    this.damage,
-    this.staminaCost,
-    this.hold,
-    this.priority,
-    this.priorityIcon,
-    this.synergy,
-    this.synergyEffects,
-    this.targets,
-    this.description,
+    required this.name,
+    required this.wikiUrl,
+    required this.type,
+    required this.classTouch,
+    required this.classIcon,
+    required this.damage,
+    required this.staminaCost,
+    required this.hold,
+    required this.priority,
+    required this.priorityIcon,
+    required this.synergy,
+    required this.synergyEffects,
+    required this.targets,
+    required this.description,
   });
 
   factory TemTemApiTechnique.fromJson(Map<String, dynamic> json) {
     return TemTemApiTechnique(
-      name: json['name'],
-      wikiUrl: json['wikiUrl'],
-      type: json['type'],
-      classTouch: json['class'],
-      classIcon: json['classIcon'],
-      damage: json['damage'],
-      staminaCost: json['staminaCost'],
-      hold: json['hold'],
-      priority: json['priority'],
-      priorityIcon: json['priorityIcon'],
-      synergy: json['synergy'],
-      synergyEffects: (json['synergyEffects'] as List)
-          .map<SynergyEffect>((item) => SynergyEffect.fromJson(item))
+      name: json['name'] as String,
+      wikiUrl: json['wikiUrl'] as String,
+      type: json['type'] as String,
+      classTouch: json['class'] as String,
+      classIcon: json['classIcon'] as String,
+      damage: json['damage'] as int,
+      staminaCost: json['staminaCost'] as int,
+      hold: json['hold'] as int,
+      priority: json['priority'] as String,
+      priorityIcon: json['priorityIcon'] as String,
+      synergy: json['synergy'] as String,
+      synergyEffects: (json['synergyEffects'] as Iterable)
+          .cast<Map<String, dynamic>>()
+          .map<SynergyEffect>(SynergyEffect.fromJson)
           .toList(),
       targets: json['targets'].toString(),
       description: json['description'].toString().replaceAll('\n', ''),