+++ lib/src/model/temtem_api_tem/temtem_api_tem.dart
@@ -5,29 +5,7 @@ part '_details.dart';
 part '_technique.dart';
 part '_evolution_node.dart';
 part '_evolution.dart';
-
-class _FreeTem {
-  final int minLevel;
-  final int maxLevel;
-  final int minPansuns;
-  final int maxPansuns;
-
-  _FreeTem({
-    required this.minLevel,
-    required this.maxLevel,
-    required this.minPansuns,
-    required this.maxPansuns,
-  });
-
-  factory _FreeTem.fromJson(Map<String, dynamic> json) {
-    return _FreeTem(
-      minLevel: json['minLevel'],
-      maxLevel: json['maxLevel'],
-      minPansuns: json['minPansuns'],
-      maxPansuns: json['maxPansuns'],
-    );
-  }
-}
+part '_free_tem.dart';
 
 class TemLocation {
   final String location;
@@ -36,7 +14,7 @@ class TemLocation {
   final String island;
   final String frequency;
   final String level;
-  final _FreeTem freeTem;
+  final FreeTem freeTem;
 
   TemLocation({
     required this.location,
@@ -50,13 +28,13 @@ class TemLocation {
 
   factory TemLocation.fromJson(Map<String, dynamic> json) {
     return TemLocation(
-      location: json['location'],
-      place: json['place'],
-      note: json['note'],
-      island: json['island'],
-      frequency: json['frequency'],
-      level: json['level'],
-      freeTem: _FreeTem.fromJson(json['freetem']),
+      location: json['location'] as String,
+      place: json['place'] as String,
+      note: json['note'] as String,
+      island: json['island'] as String,
+      frequency: json['frequency'] as String,
+      level: json['level'] as String,
+      freeTem: FreeTem.fromJson(json['freetem'] as Map<String, dynamic>),
     );
   }
 }
@@ -93,12 +71,12 @@ class TemTemApiTem {
   final String? lumaPortraitWikiUrl;
   final List<String> traits;
   final Details details;
-  final List<Technique> techniques;
+  final List<Technique>? techniques;
   final List<String> trivia;
   final Evolution evolution;
   final String wikiPortraitUrlLarge;
   final String? lumaWikiPortraitUrlLarge;
-  final List<TemLocation> locations;
+  final List<TemLocation>? locations;
   final String icon;
   final String lumaIcon;
   final GenderRatio genderRatio;
@@ -149,39 +127,43 @@ class TemTemApiTem {
   });
 
   factory TemTemApiTem.fromJson(Map<String, dynamic> json) {
+    final techniques =
+        (json['techniques'] as Iterable?)?.cast<Map<String, dynamic>>();
+    final locations =
+        (json['locations'] as Iterable?)?.cast<Map<String, dynamic>>();
+
     return TemTemApiTem(
-      number: json['number'],
-      name: json['name'],
-      types: List<String>.from(json['types']),
-      portraitWikiUrl: json['portraitWikiUrl'],
-      lumaPortraitWikiUrl: json['lumaPortraitWikiUrl'],
-      wikiUrl: json['wikiUrl'],
-      stats: Stats.fromJson(json['stats']),
-      traits: List<String>.from(json['traits']),
-      details: Details.fromJson(json['details']),
-      techniques: Iterable<Technique>.generate(json['techniques'].length,
-          (index) => Technique.fromJson(json['techniques'][index])).toList(),
-      trivia: List<String>.from(json['trivia']),
-      evolution: Evolution.fromJson(json['evolution']),
-      wikiPortraitUrlLarge: json['wikiPortraitUrlLarge'],
-      lumaWikiPortraitUrlLarge: json['lumaWikiPortraitUrlLarge'],
-      locations: Iterable<TemLocation>.generate(json['locations'].length,
-          (index) => TemLocation.fromJson(json['locations'][index])).toList(),
-      icon: json['icon'],
-      lumaIcon: json['lumaIcon'],
-      genderRatio: GenderRatio.fromJson(json['genderRatio']),
+      number: json['number'] as int,
+      name: json['name'] as String,
+      types: List<String>.from(json['types'] as Iterable),
+      portraitWikiUrl: json['portraitWikiUrl'] as String,
+      lumaPortraitWikiUrl: json['lumaPortraitWikiUrl'] as String?,
+      wikiUrl: json['wikiUrl'] as String,
+      stats: Stats.fromJson(json['stats'] as Map<String, dynamic>),
+      traits: List<String>.from(json['traits'] as Iterable),
+      details: Details.fromJson(json['details'] as Map<String, dynamic>),
+      techniques: techniques?.map(Technique.fromJson).toList(),
+      trivia: List<String>.from(json['trivia'] as Iterable),
+      evolution: Evolution.fromJson(json['evolution'] as Map<String, dynamic>),
+      wikiPortraitUrlLarge: json['wikiPortraitUrlLarge'] as String,
+      lumaWikiPortraitUrlLarge: json['lumaWikiPortraitUrlLarge'] as String?,
+      locations: locations?.map(TemLocation.fromJson).toList(),
+      icon: json['icon'] as String,
+      lumaIcon: json['lumaIcon'] as String,
+      genderRatio:
+          GenderRatio.fromJson(json['genderRatio'] as Map<String, dynamic>),
       catchRate: json['catchRate'] as int,
       hatchMins: json['hatchMins'] as int,
-      tvYields: Stats.fromJson(json['tvYields']),
-      gameDescription: json['gameDescription'],
-      wikiRenderStaticUrl: json['wikiRenderStaticUrl'],
-      renderStaticImage: json['renderStaticImage'],
-      wikiRenderStaticLumaUrl: json['wikiRenderStaticLumaUrl'],
-      wikiRenderAnimatedUrl: json['wikiRenderAnimatedUrl'],
-      renderAnimatedImage: json['renderAnimatedImage'],
-      wikiRenderAnimatedLumaUrl: json['wikiRenderAnimatedLumaUrl'],
-      renderStaticLumaImage: json['renderStaticLumaImage'],
-      renderAnimatedLumaImage: json['renderAnimatedLumaImage'],
+      tvYields: Stats.fromJson(json['tvYields'] as Map<String, dynamic>),
+      gameDescription: json['gameDescription'] as String,
+      wikiRenderStaticUrl: json['wikiRenderStaticUrl'] as String?,
+      renderStaticImage: json['renderStaticImage'] as String?,
+      wikiRenderStaticLumaUrl: json['wikiRenderStaticLumaUrl'] as String?,
+      wikiRenderAnimatedUrl: json['wikiRenderAnimatedUrl'] as String?,
+      renderAnimatedImage: json['renderAnimatedImage'] as String?,
+      wikiRenderAnimatedLumaUrl: json['wikiRenderAnimatedLumaUrl'] as String?,
+      renderStaticLumaImage: json['renderStaticLumaImage'] as String?,
+      renderAnimatedLumaImage: json['renderAnimatedLumaImage'] as String?,
     );
   }
 }