+++ lib/src/model/temtem_api_traits.dart
@@ -3,13 +3,17 @@ class TemTemApiTraits {
   final String wikiUrl;
   final String description;
 
-  TemTemApiTraits({this.name, this.wikiUrl, this.description});
+  TemTemApiTraits({
+    required this.name,
+    required this.wikiUrl,
+    required this.description,
+  });
 
   factory TemTemApiTraits.fromJson(Map<String, dynamic> json) {
     return TemTemApiTraits(
-      name: json['name'],
-      wikiUrl: json['wikiUrl'],
-      description: json['description'],
+      name: json['name'] as String,
+      wikiUrl: json['wikiUrl'] as String,
+      description: json['description'] as String,
     );
   }
 }