+++ lib/src/model/temtem_api_weakness.dart
@@ -1,7 +1,7 @@
 class TemTemApiWeakness {
   final Map<String, dynamic> weaknesses;
 
-  TemTemApiWeakness({this.weaknesses});
+  TemTemApiWeakness({required this.weaknesses});
 
   factory TemTemApiWeakness.fromJson(Map<String, dynamic> json) =>
       TemTemApiWeakness(weaknesses: Map.from(json));