+++ lib/src/model/temtem_api_weakness_calc.dart
@@ -5,17 +5,17 @@ class TemTemApiWeaknessCalc {
   final double result;
 
   TemTemApiWeaknessCalc({
-    this.attacking,
-    this.defending,
-    this.modifiers,
-    this.result,
+    required this.attacking,
+    required this.defending,
+    required this.modifiers,
+    required this.result,
   });
 
   factory TemTemApiWeaknessCalc.fromJson(Map<String, dynamic> json) =>
       TemTemApiWeaknessCalc(
-        attacking: json['attacking'],
-        defending: List<String>.from(json['defending']),
-        modifiers: List<double>.from(json['modifiers']),
-        result: json['result'],
+        attacking: json['attacking'] as String,
+        defending: List<String>.from(json['defending'] as Iterable),
+        modifiers: List<double>.from(json['modifiers'] as Iterable),
+        result: json['result'] as double,
       );
 }